The number of walkers is restored when the samplesperthread is small.
git-svn-id: https://subversion.assembla.com/svn/qmcdev/trunk@5906 e5b18d87-469d-4833-9cc0-8cdfa06e9491

+++ src/QMCDrivers/QMCLinearOptimize.cpp
@@ -122,7 +122,10 @@ void QMCLinearOptimize::finish()
   optTarget->reportParameters();
   int nw_removed=W.getActiveWalkers()-NumOfVMCWalkers;
   app_log() << "   Restore the number of walkers to " << NumOfVMCWalkers << ", removing " << nw_removed << " walkers." <<endl;
-  W.destroyWalkers(nw_removed);
+  if(nw_removed>0)
+    W.destroyWalkers(nw_removed);
+  else
+    W.createWalkers(-nw_removed);
   app_log() << "</opt>" << endl;
   app_log() << "</optimization-report>" << endl;
 }
